refactor: delete bonus_colour.c/h
